requirements.txt
streamlit>=1.32.0
plotly>=5.19.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
statsmodels>=0.14.0
openpyxl>=3.1.0
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-storage>=2.24.0
pyarrow>=14.0.0
db-dtypes>=1.2.0
xgboost>=2.0.0
